Added stepper indicator
