--- app/signup/steps/chooseCodeLanguage.tsx
@@ -1,8 +1,0 @@
-"use client";
-
-import React from "react";
-
-import { motion } from "framer-motion";
-import Image from "next/image";
-
-/// choose code language with the flags in a carousel
