+++ app/sitemap.ts
@@ -30,7 +30,7 @@ export default function sitemap(): MetadataRoute.Sitemap {
         url: 'https://angova.io/signup',
         lastModified: new Date(),
         changeFrequency: 'weekly',
-        priority: 0.5,
+        priority: 0.3,
       }
   ]
 }
